Refactor name and title components
