+++ src/app/components/SplashScreen.js
@@ -14,7 +14,7 @@ export default function SplashScreen() {
     ,[]);
 
     return (
-        <div id="splash-screen" className={`fixed z-[100]  items-center flex flex-col w-screen h-screen bg-black ${!isSplashScreenVisible&&"hidden"}`}>
+        <div id="splash-screen" className={`fixed z-[100]  items-center flex flex-col w-screen h-screen bg-black ${!isSplashScreenVisible&&"fade"}`}>
             <div className="flex flex-row items-center justify-center h-5/6">
                     <LogoLined className="logo md:h-24 h-32" />
             </div>
